requirements.txt
selenium==4.20.0
pytest==8.2.1
pytest-html==4.1.1
pytest-xdist==3.6.1
allure-pytest==2.13.2
webdriver-manager==4.0.1
